fix --directory option
