+++ lib/node-pre-gyp.js
@@ -58,19 +58,22 @@ exports.find = require('./pre-binding').find;
 //
 // that's why "find()" must pass the path to package.json.
 //
-function Run(package_json_path = './package.json') {
-  this.package_json = JSON.parse(fs.readFileSync(package_json_path));
-
+function Run({ package_json_path = './package.json', argv }) {
+  this.package_json_path = package_json_path;
   this.commands = {};
 
   const self = this;
   cli_commands.forEach((command) => {
-    self.commands[command] = function(argv, callback) {
-      log.verbose('command', command, argv);
-      return require('./' + command)(self, argv, callback);
+    self.commands[command] = function(argvx, callback) {
+      log.verbose('command', command, argvx);
+      return require('./' + command)(self, argvx, callback);
     };
   });
-  // the destination for the binaries has not yet been set.
+
+  this.parseArgv(argv);
+
+  // this is set to true after the binary.host property was set to
+  // either staging_host or production_host.
   this.binaryHostSet = false;
 }
 inherits(Run, EE);
@@ -117,8 +120,8 @@ proto.shorthands = {
 proto.aliases = aliases;
 
 /**
- * Parses the given argv array and sets the 'opts',
- * 'argv' and 'command' properties.
+ * Parses the given argv array and sets the 'opts', 'argv',
+ * 'command', and 'package_json' properties.
  */
 
 proto.parseArgv = function parseOpts(argv) {
@@ -127,16 +130,16 @@ proto.parseArgv = function parseOpts(argv) {
   const commands = this.todo = [];
 
   // create a copy of the argv array with aliases mapped
-  argv = this.argv.map(function(arg) {
+  argv = this.argv.map((arg) => {
     // is this an alias?
     if (arg in this.aliases) {
       arg = this.aliases[arg];
     }
     return arg;
-  }, this);
+  });
 
   // process the mapped args into "command" objects ("name" and "args" props)
-  argv.slice().forEach(function(arg) {
+  argv.slice().forEach((arg) => {
     if (arg in this.commands) {
       const args = argv.splice(0, argv.indexOf(arg));
       argv.shift();
@@ -145,21 +148,27 @@ proto.parseArgv = function parseOpts(argv) {
       }
       commands.push({ name: arg, args: [] });
     }
-  }, this);
+  });
   if (commands.length > 0) {
     commands[commands.length - 1].args = argv.splice(0);
   }
 
-  // expand commands entries for multiple napi builds
-  let dir = this.opts.directory;
-  if (dir == null) dir = process.cwd();
-  const package_json = JSON.parse(fs.readFileSync(path.join(dir, 'package.json')));
 
-  this.todo = napi.expand_commands(package_json, this.opts, commands);
+  // if a directory was specified package.json is assumed to be relative
+  // to it.
+  let package_json_path = this.package_json_path;
+  if (this.opts.directory) {
+    package_json_path = path.join(this.opts.directory, package_json_path);
+  }
+
+  this.package_json = JSON.parse(fs.readFileSync(package_json_path));
+
+  // expand commands entries for multiple napi builds
+  this.todo = napi.expand_commands(this.package_json, this.opts, commands);
 
   // support for inheriting config env variables from npm
   const npm_config_prefix = 'npm_config_';
-  Object.keys(process.env).forEach(function(name) {
+  Object.keys(process.env).forEach((name) => {
     if (name.indexOf(npm_config_prefix) !== 0) return;
     const val = process.env[name];
     if (name === npm_config_prefix + 'loglevel') {
@@ -182,7 +191,7 @@ proto.parseArgv = function parseOpts(argv) {
         this.opts[name] = val;
       }
     }
-  }, this);
+  });
 
   if (this.opts.loglevel) {
     log.level = this.opts.loglevel;
