+++ lib/pre-binding.js
@@ -18,8 +18,7 @@ exports.find = function(package_json_path, opts) {
   if (!existsSync(package_json_path)) {
     throw new Error(package_json_path + 'does not exist');
   }
-  const prog = new npg.Run(package_json_path);
-  prog.parseArgv(process.argv);
+  const prog = new npg.Run({ package_json_path, argv: process.argv });
   prog.setBinaryHostProperty();
   const package_json = prog.package_json;
